requirements.txt
open3d==0.18.0
wandb
trimesh==4.0.10
pytorch_msssim
torchmetrics
tqdm
imageio
opencv-python
plyfile
roma
einops==0.8.0
numpy==1.26.4
PyQt5==5.15.11
matplotlib==3.5.1
evo==1.11.0
python-pycg
einops
git+https://github.com/VladimirYugay/simple-knn.git@c7e51a06a4cd84c25e769fee29ab391fe5d5ff8d
git+https://github.com/eriksandstroem/evaluate_3d_reconstruction_lib.git@9b3cc08be5440db9c375cc21e3bd65bb4a337db7
# git+https://github.com/VladimirYugay/gaussian_rasterizer.git@9c40173fcc8d9b16778a1a8040295bc2f9ebf129
# git+https://github.com/rmurai0610/diff-gaussian-rasterization-w-pose.git@43e21bff91cd24986ee3dd52fe0bb06952e50ec7